requirements.txt
python-dotenv
pdfplumber
langchain
langchain-google-genai
google-generativeai
streamlit